Allow lazy loading extensions
This should fix the search extension. With the following error:

```
Failed to load resource: the server responded with a status of 404 (Not Found) on /assets/ace/ext-searchbox.js
```

In order to make it work the search module should be loaded with the
proper asset_path reference, just like the workers and modes.